+++ R/CohortCharacterizationDiagnostics.R
@@ -82,11 +82,11 @@ getCohortCharacteristics <- function(connectionDetails = NULL,
     covariates <- featureExtractionOutput$covariates %>%
       dplyr::rename("cohortId" = "cohortDefinitionId") %>%
       dplyr::left_join(populationSize, by = "cohortId", copy = TRUE) %>%
-      dplyr::mutate(p = sumValue / populationSize)
+      dplyr::mutate(p = .data$sumValue / populationSize)
 
     if (nrow(covariates %>%
-               dplyr::filter(p > 1) %>%
-               dplyr::collect()) > 0) {
+      dplyr::filter(.data$p > 1) %>%
+      dplyr::collect()) > 0) {
       stop(
         paste0(
           "During characterization, population size (denominator) was found to be smaller than features Value (numerator).",
@@ -96,10 +96,10 @@ getCohortCharacteristics <- function(connectionDetails = NULL,
     }
 
     covariates <- covariates %>%
-      dplyr::mutate(sd = sqrt(p * (1 - p))) %>%
-      dplyr::select(-p) %>%
+      dplyr::mutate(sd = sqrt(.data$p * (1 - .data$p))) %>%
+      dplyr::select(-"p") %>%
       dplyr::rename("mean" = "averageValue") %>%
-      dplyr::select(-populationSize)
+      dplyr::select(-"populationSize")
 
     if (FeatureExtraction::isTemporalCovariateData(featureExtractionOutput)) {
       covariates <- covariates %>%
@@ -113,13 +113,13 @@ getCohortCharacteristics <- function(connectionDetails = NULL,
         )
 
       tidNaCount <- covariates %>%
-        dplyr::filter(is.na(timeId)) %>%
+        dplyr::filter(is.na(.data$timeId)) %>%
         dplyr::count() %>%
         dplyr::pull()
 
       if (tidNaCount > 0) {
         covariates <- covariates %>%
-          dplyr::mutate(timeId = if_else(is.na(.data$timeId), -1, .data$timeId))
+          dplyr::mutate(timeId = dplyr::if_else(is.na(.data$timeId), -1, .data$timeId))
       }
     } else {
       covariates <- covariates %>%
@@ -144,9 +144,9 @@ getCohortCharacteristics <- function(connectionDetails = NULL,
     dplyr::pull(dplyr::count(featureExtractionOutput$covariatesContinuous)) > 0) {
     covariates <- featureExtractionOutput$covariatesContinuous %>%
       dplyr::rename(
-        mean = averageValue,
-        sd = standardDeviation,
-        cohortId = cohortDefinitionId
+        mean = "averageValue",
+        sd = "standardDeviation",
+        cohortId = "cohortDefinitionId"
       )
     covariatesContinuous <- covariates
     if (FeatureExtraction::isTemporalCovariateData(featureExtractionOutput)) {
@@ -162,13 +162,13 @@ getCohortCharacteristics <- function(connectionDetails = NULL,
         )
 
       tidNaCount <- covariates %>%
-        dplyr::filter(is.na(timeId)) %>%
+        dplyr::filter(is.na(.data$timeId)) %>%
         dplyr::count() %>%
         dplyr::pull()
 
       if (tidNaCount > 0) {
         covariates <- covariates %>%
-          dplyr::mutate(timeId = if_else(is.na(.data$timeId), -1, .data$timeId))
+          dplyr::mutate(timeId = dplyr::if_else(is.na(.data$timeId), -1, .data$timeId))
       }
     } else {
       covariates <- covariates %>%
@@ -235,7 +235,7 @@ executeCohortCharacterization <- function(connection,
   startCohortCharacterization <- Sys.time()
   subset <- subsetToRequiredCohorts(
     cohorts = cohorts %>%
-      dplyr::filter(cohortId %in% instantiatedCohorts),
+      dplyr::filter(.data$cohortId %in% instantiatedCohorts),
     task = task,
     incremental = incremental,
     recordKeepingFile = recordKeepingFile
@@ -286,7 +286,7 @@ executeCohortCharacterization <- function(connection,
           tempEmulationSchema = tempEmulationSchema,
           cohortDatabaseSchema = cohortDatabaseSchema,
           cohortTable = cohortTable,
-          cohortIds = subset[start:end,]$cohortId,
+          cohortIds = subset[start:end, ]$cohortId,
           covariateSettings = covariateSettings,
           cdmVersion = cdmVersion,
           exportFolder = exportFolder
@@ -308,9 +308,9 @@ executeCohortCharacterization <- function(connection,
       )
 
       recordTasksDone(
-        cohortId = subset[start:end,]$cohortId,
+        cohortId = subset[start:end, ]$cohortId,
         task = task,
-        checksum = subset[start:end,]$checksum,
+        checksum = subset[start:end, ]$checksum,
         recordKeepingFile = recordKeepingFile,
         incremental = incremental
       )