+++ NEWS.md
@@ -1,3 +1,22 @@
+CohortDiagnostics 3.3.0
+=======================
+
+Changes: 
+
+1. Resolved issues with package creating build errors
+
+2. Added function to make deployment to posit connect servers easier
+
+3. Added ability to use CohortGenerator sample functionality to executeDiagnostics which speeds up execution for very
+large cohort definitions
+
+Bug fix:
+
+1. Fixes to loading shiny app in OHDSI shiny modules
+
+2. Bug fixes to prevent null value issues with included source concept diagnostics
+
+
 CohortDiagnostics 3.2.4
 =======================
 Bug Fix: